refactor: split map generation into passes
