+++ src/main/java/com/mapbuilder/mapbuilder/core/map/generator/MapGenerationPass.java
@@ -1,0 +1,6 @@
+package com.mapbuilder.mapbuilder.core.map.generator;
+import com.mapbuilder.mapbuilder.core.map.MapGrid;
+
+public interface MapGenerationPass {
+    void execute(MapGrid grid, GenerationParameters params);
+}
